Fixes GitHub Metric Graphs
