+++ Frontend/nextjs-dashboard/pages/metrics.tsx
@@ -4,6 +4,7 @@ import JoyAvgChart from '@/app/ui/metrics/joy-avg-chart'
 import JoyStudentRatingGraph from '@/app/ui/metrics/joy-student-rating-graph'
 import TeamVelocityGraph from '@/app/ui/metrics/team-velocity-graph'
 import TeamVelocityInput from '@/app/ui/metrics/team-velocity-input'
+import GitHubAppAuthorizationDialog from '@/app/ui/metrics/github-app-authorization-dialog'
 import '@/app/ui/stylesheets/metrics.css'
 import { redirect } from 'next/dist/server/api-utils'
 import { useEffect } from 'react'
@@ -18,53 +19,6 @@ async function closeDialogs()
 }
 
 export default function Metrics() {
-
-    async function redirectToGitHubAppAuthorization()
-    {
-        fetch("http://127.0.0.1:3001/auth/get-github-app-client-id",
-            {
-                method: 'GET'
-            }
-        ).then(response => {
-            if (!response.ok) {
-                if (response.status === 404) {
-                    return {
-                        text: "Server not found!",
-                        status: "danger"
-                    };
-                }
-      
-                return response.text().then(text => {
-                    return {
-                        text: text,
-                        status: "danger"
-                    };
-                })
-            }
-            return response.text().then(text => {
-                return {
-                    text: text,
-                    status: "success"
-                };
-            })
-        }).then(resp => {
-            console.log("result:", resp);
-            try {
-                let r = JSON.parse(resp.text)
-                if(!r['approved'])
-                    throw 'idek bro'
-                console.log(r)
-                const client_id = r['clientId']
-                window.location.assign('https://github.com/login/oauth/authorize?client_id='+client_id)
-                
-            } catch(err) {
-                console.log(err)
-            }
-        }).catch((error) => {
-            console.log(error)
-        })
-       
-    }
 
     async function openJoyRatingDialog() 
     {
@@ -82,13 +36,13 @@ export default function Metrics() {
             <div>
                 <button onClick={openJoyRatingDialog}>Open Joy Rating Dialog</button>
                 <button onClick={openTeamVelocityDialog}>Open Team Velocity Dialog</button>
-                <button onClick={redirectToGitHubAppAuthorization}>GitHub App Authorization</button>
             </div>
             <div>
                 <ContributionsGraph />
                 <JoyAvgChart />
                 <JoyStudentRatingGraph />
                 <TeamVelocityGraph />
+                <GitHubAppAuthorizationDialog />
             </div>
             <div id="metrics-dialog-backdrop" onClick={closeDialogs}></div>
             <TeamVelocityInput />
